+++ Online Exam 1/VersionC/versionC-Ashen.c
