+++ Online Exam 1/VersionC/versionC-Thushara.c
